rename index.js to full.js and add formating config to faciltate merging
